Converted Bee Logo into an Eagle Part
Here's a library containing a single device which has 1 symbol and 1 package.
The symbol is a simple rectangle that shows up in the schematic and says "Bee Logo".
Adding the device in the schematic editor will create a copy of the image in the schematic which can be grabbed by the Anchor at the lower left corner and easily moved around for placement on the PCB as a unit.
The Bee Logo itself is in the TPlace layer (as was the original Bee Logo). The anchor for dragging it around is in the Reference layer so that the anchor doesn't get printed on the actual silk screen.
Mirroring the logo will automatically move it to the BPlace layer as well.